Migrated all diagnostics over to an `ErrorHandling` type
